changed chunk system, added terrainheight cache, and bug fixes

+++ game 2/src/entities/Entitys.zig
@@ -10,13 +10,13 @@ pub const EntityUUID = struct {
 };
 
 pub const Player = struct {
-    pos: @Vector(3, f32),
+    pos: @Vector(3, f64),
     pitch: f32,
     yaw: f32,
     roll: f32,
     speed: @Vector(3, f32),
-    cameraUp: @Vector(3, f32),
-    cameraFront: @Vector(3, f32),
+    cameraUp: @Vector(3, f64),
+    cameraFront: @Vector(3, f64),
     GenDistance: [3]u32,
     LoadDistance: [3]u32,
     MeshDistance: [3]u32,
